+++ src/main/java/com/dalbit/member/vo/procedure/P_ExchangeVo.java
@@ -3,11 +3,11 @@ package com.dalbit.member.vo.procedure;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.validation.constraints.Min;
+import java.math.BigDecimal;
 
 @Getter @Setter
 public class P_ExchangeVo {
 
     private String mem_no;
-    private int byeol;
+    private BigDecimal byeol;
 }
