+++ src/main/java/com/dalbit/member/vo/request/ExchangeReApplyVo.java
@@ -4,15 +4,14 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
+import java.math.BigDecimal;
 
 @Getter @Setter
 public class ExchangeReApplyVo {
 
     private int exchangeIdx;
     @Min(message = "{\"ko_KR\" : \"별을\"}", value = 570)
-    private int byeol;
+    private BigDecimal byeol;
 
     private String accountName;
     private String accountNo;
