+++ src/main/java/com/dalbit/member/vo/request/ExchangeVo.java
@@ -4,10 +4,11 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.validation.constraints.Min;
+import java.math.BigDecimal;
 
 @Getter @Setter
 public class ExchangeVo {
 
     @Min(message = "{\"ko_KR\" : \"별을\"}", value = 570)
-    private int byeol;
+    private BigDecimal byeol;
 }
